refactor(css): add css folder, put all css files in it and fix references to them
