+++ front-end/src/css/VerCavalo.css
